Initial commit
GameController is manages the screens and completed.

Main Menu is works but not designed.

Stage:
Reads the first level and run the game.
There is a collision between the player to the other objects excluding the gifts.
There is a collision between the stupid and random enemy to the relevant objects.

Stage construction does not work.